# docs/contracts/api/openapi.yaml
openapi: 3.1.0
info:
  title: LexFrame Stage 8 API
  version: 0.8.0
  description: Canonical HTTP contract for the LexFrame stage 8 execution core, embedded builder, approval inbox, artifact handling, delivery and notifications.
servers:
  - url: http://localhost:3100
paths:
...
  /legal-search/query:
    post:
      operationId: queryLegalSearch
      tags: [LegalResearch]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LegalSearchQuery"
      responses:
        "200":
          description: Search response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LegalSearchResponse"
...
  /workflow-runtime/legal-search/execute:
    post:
      operationId: executeWorkflowRuntimeLegalSearch
      tags: [WorkflowRuntime]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LegalSearchQuery"
      responses:
        "200":
          description: Runtime search response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LegalSearchResponse"
...
  /notifications:
    get:
      operationId: listNotifications
      tags: [Notifications]
      parameters:
        - in: query
          name: cursor
          required: false
          schema:
            type: string
        - in: query
          name: limit
          required: false
          schema:
            type: integer
        - in: query
          name: status
          required: false
          schema:
            type: string
            enum: [all, unread, read]
        - in: query
          name: type
          required: false
          schema:
            type: string
      responses:
        "200":
          description: User-scoped notifications inbox
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationListResponse"
  /notifications/{id}/read:
    post:
      operationId: markNotificationRead
      tags: [Notifications]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Notification marked as read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationSummary"
...
  /devices/register:
    post:
      operationId: registerDevice
      tags: [Notifications]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeviceRegistrationRequest"
      responses:
        "200":
          description: Device token registered for notifications
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegisteredDevice"
  /devices/{id}:
    delete:
      operationId: removeDevice
      tags: [Notifications]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Device token removed
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status:
                    type: string
                    enum: [removed]
...
components:
...
  schemas:
...
    LegalSourceProviderSummary:
      type: object
      required:
        [id, code, name, providerType, jurisdiction, accessMode, isEnabled]
      properties:
        id:
          type: string
        code:
          type: string
        name:
          type: string
        providerType:
          type: string
          enum:
            [
              user_upload,
              workspace_private,
              product_curated,
              official_feed,
              external_paid,
              manual_import,
            ]
        jurisdiction:
          type: [string, "null"]
        accessMode:
          type: string
          enum: [file_upload, api_import, seeded, manual]
        isEnabled:
          type: boolean
...
    LegalChunkSummary:
      type: object
      required:
        [
          id,
          sourceId,
          documentVersionId,
          chunkNo,
          chunkType,
          text,
          textHash,
          pageFrom,
          pageTo,
          charStart,
          charEnd,
          metadata,
          securityScope,
          embeddingModel,
          embeddingHash,
          indexedAt,
        ]
      properties:
        id:
          type: string
        sourceId:
          type: string
        documentVersionId:
          type: string
        chunkNo:
          type: integer
        chunkType:
          type: string
          enum:
            [
              facts,
              claims,
              court_reasoning,
              holding,
              procedural_history,
              citations,
              operative_part,
              unknown,
            ]
        text:
          type: string
        textHash:
          type: string
        pageFrom:
          type: [integer, "null"]
        pageTo:
          type: [integer, "null"]
        charStart:
          type: [integer, "null"]
        charEnd:
          type: [integer, "null"]
        metadata:
          type: object
          additionalProperties: true
        securityScope:
          type: string
          enum:
            [
              public,
              product_private,
              workspace_private,
              user_private,
              restricted_provider,
            ]
        embeddingModel:
          type: [string, "null"]
        embeddingHash:
          type: [string, "null"]
        indexedAt:
          type: [string, "null"]
          format: date-time
...
    LegalSourceSummary:
      type: object
      required:
        [
          id,
          workspaceId,
          documentId,
          provider,
          sourceType,
          jurisdiction,
          title,
          canonicalUrl,
          externalId,
          licenseStatus,
          visibility,
          classification,
          status,
          ownerWorkspaceId,
          ownerUserId,
          court,
          caseNumber,
          decisionDate,
          hasEmbeddings,
          indexedAt,
          lastUsedAt,
          createdAt,
          updatedAt,
        ]
      properties:
        id:
          type: string
        workspaceId:
          type: [string, "null"]
        documentId:
          type: [string, "null"]
        provider:
          $ref: "#/components/schemas/LegalSourceProviderSummary"
        sourceType:
          type: string
          enum:
            [
              court_decision,
              statute,
              regulation,
              contract_template,
              user_document,
              internal_memo,
              analysis_result,
            ]
        jurisdiction:
          type: [string, "null"]
        title:
          type: string
        canonicalUrl:
          type: [string, "null"]
          format: uri
        externalId:
          type: [string, "null"]
        licenseStatus:
          type: string
          enum: [allowed, restricted, unknown, requires_contract, forbidden]
        visibility:
          type: string
          enum:
            [
              public,
              product_private,
              workspace_private,
              user_private,
              restricted_provider,
            ]
        classification:
          type: string
          enum:
            [
              public,
              internal,
              confidential,
              legal_secret,
              personal_data,
              client_material,
            ]
        status:
          type: string
          enum:
            [
              draft,
              pending_processing,
              processed,
              indexed,
              index_failed,
              deprecated,
              archived,
              deleted,
            ]
        ownerWorkspaceId:
          type: [string, "null"]
        ownerUserId:
          type: [string, "null"]
        court:
          type: [string, "null"]
        caseNumber:
          type: [string, "null"]
        decisionDate:
          type: [string, "null"]
        hasEmbeddings:
          type: boolean
        indexedAt:
          type: [string, "null"]
          format: date-time
        lastUsedAt:
          type: [string, "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    LegalSearchFilters:
      type: object
      properties:
        sourceType:
          type: array
          items:
            type: string
            enum:
              [
                court_decision,
                statute,
                regulation,
                contract_template,
                user_document,
                internal_memo,
                analysis_result,
              ]
        visibility:
          type: array
          items:
            type: string
            enum:
              [
                public,
                product_private,
                workspace_private,
                user_private,
                restricted_provider,
              ]
        court:
          type: array
          items:
            type: string
        dateFrom:
          type: string
          format: date
        dateTo:
          type: string
          format: date
        category:
          type: array
          items:
            type: string
        workspaceId:
          type: string
        caseNumber:
          type: string
    LegalSearchQuery:
      type: object
      required: [query, mode]
      properties:
        query:
          type: string
        mode:
          type: string
          enum: [keyword, semantic, hybrid]
        filters:
          $ref: "#/components/schemas/LegalSearchFilters"
        limit:
          type: integer
        offset:
          type: integer
        selectedSourceIds:
          type: array
          items:
            type: string
    LegalSearchFacetBucket:
      type: object
      required: [value, count]
      properties:
        value:
          type: string
        count:
          type: integer
    LegalSearchFacet:
      type: object
      required: [name, label, buckets]
      properties:
        name:
          type: string
        label:
          type: string
        buckets:
          type: array
          items:
            $ref: "#/components/schemas/LegalSearchFacetBucket"
    LegalCitation:
      type: object
      required:
        [
          citationId,
          sourceId,
          chunkId,
          documentVersionId,
          title,
          quote,
          pageFrom,
          pageTo,
          court,
          caseNumber,
          decisionDate,
          score,
        ]
      properties:
        citationId:
          type: string
        sourceId:
          type: string
        chunkId:
          type: string
        documentVersionId:
          type: string
        title:
          type: string
        quote:
          type: string
        pageFrom:
          type: [integer, "null"]
        pageTo:
          type: [integer, "null"]
        court:
          type: [string, "null"]
        caseNumber:
          type: [string, "null"]
        decisionDate:
          type: [string, "null"]
        score:
          type: number
    LegalSearchResult:
      type: object
      required:
        [
          rank,
          score,
          scoreComponents,
          source,
          chunk,
          snippet,
          highlights,
          citation,
        ]
      properties:
        rank:
          type: integer
        score:
          type: number
        scoreComponents:
          type: object
          required: [lexical, semantic, combined]
          properties:
            lexical:
              type: number
            semantic:
              type: number
            combined:
              type: number
        source:
          $ref: "#/components/schemas/LegalSourceSummary"
        chunk:
          $ref: "#/components/schemas/LegalChunkSummary"
        snippet:
          type: string
        highlights:
          type: array
          items:
            type: string
        citation:
          $ref: "#/components/schemas/LegalCitation"
    LegalSearchResponse:
      type: object
      required: [mode, total, facets, results, debug]
      properties:
        mode:
          type: string
          enum: [keyword, semantic, hybrid]
        total:
          type: integer
        facets:
          type: array
          items:
            $ref: "#/components/schemas/LegalSearchFacet"
        results:
          type: array
          items:
            $ref: "#/components/schemas/LegalSearchResult"
        debug:
          type: object
          required: [indexAlias, normalized, aclApplied]
          properties:
            indexAlias:
              type: string
            normalized:
              type: boolean
            aclApplied:
              type: boolean
...
    Stage8Object:
      type: object
      additionalProperties: true
...
    NotificationSummary:
      allOf:
        - $ref: "#/components/schemas/Stage8Object"
...
    NotificationListResponse:
      allOf:
        - $ref: "#/components/schemas/Stage8Object"
    RegisteredDevice:
      allOf:
        - $ref: "#/components/schemas/Stage8Object"
    DeviceRegistrationRequest:
      allOf:
        - $ref: "#/components/schemas/Stage8Object"
...